[FE] Feat: 매거진 상세 페이지 컴포넌트 id 적용 #358
- 매거진 상세 페이지 component id / data 적용
